Start adding database tests. Fix an error in structure definitions.
Start adding tests + framework for the storage implementation. This
revealed that node and leaf objects did not have a TreeID, which we need
to store them. I've added this to NodeID and Leaf, though it arguably
could have been part of Node too. Fix up all corresponding references
and tests.