Fix further hash confusion between raw and merkle hash (#203)
* Fix further hash confusion

The LeafData has must be the raw hash as we use it for deduping. The
sequenced data needs both raw and merkle hashes as it must both link to
the LeafData and be queried by Merkle Hash. Renamed db columns to make
it clearer which is which.

golint fix.

* Rename LeafHash to MerkleLeafHash

* Do a lot of renaming to clarify hashes and align code + db columns

Further code / db alignment of hash naming

* Fix a missing field name update

* Rename leaf proto fields to match structures and database names

Slightly naughty but nobody should be using them yet.

* TODO splitting leaf value hash in leaf struct