Rewrite the skip level code in merkle path calculations (#249)
* Rewrite the skip level code in merkle path calculations

Now is explicitly based on the tree shape at the snapshot size and
hopefully easier to read. Should be read in conjunction with
compact_merkle_tree, which is where node writes come from. Add a few
more tests, more should be written.

* lint fix

* Add some more path tests

Tests all paths on trees up to size 99 do not give an error for
inclusion + consistency paths.

* Review fixes.

* Fix an incorrect log level

* round 2 of reviews.