Remove the code for signing log roots if the last one is too old. (#266)
* Remove the code for signing log roots if the last one is too old.

This is leftover from when we thought app layers wouldn't have their own
storage. I don't think it's being used and probably doesn't do what we
need anyway.

* Create a root in sequencing if none already exists.

Could be better to move this to log create API when we write it. Then we
can assume a root exists and raise an error if it doesn't. TODO added.