Adding an ExtensionRegistry interface to represent Trillian extensions (#271)
Adding an extension.Registry interface to represent Trillian extensions

The extension.Registry contains methods to acquire both Log and MapStorages. In
the future it may be extended to contain other pluggable components, such as
AdminStorage, KeyStorage and so on.

Both log and map servers have been changed to use the new abstraction.