Fix GROUP BY for map on newer mysql versions
And update the doc. I think this is correct because we don't expect to
get multiple rows with the same keyhash and value in a consistent map
tree hence none are collapsed to a group in the result set and if I wave
my hands and say this very loudly then people will probably believe me.