Move treeID parameter from GetStorage methods to tx.Begin / tx.Snapshot
Currently we have a storage-per-tree abstraction. This replaces that abstraction
with transaction-per-tree. A future PR will allow us to create transactions not
tied to a particular tree. An example would be not having to rely on tree #0
being present during server startup.

A few other changes made here are:
* Make sequencer read the logID from the context (sequencer instances are not
  per-tree anymore)
* Make sequencer tests use an explicit logID (the ID was buried in mocks, so it
  didn't need to / wasn't set)
* Use spread out ports in ct_integration_test (fixes some flakiness I've been
  experiencing locally).
* Added a few tests for Begin/Snapshot (some to expanded in the future)
* A few cosmetic changes (moving "if err != nil" closer to the declaration of
  err, mostly)

I'm happy to make follow up changes, but I'd like to have them addressed in
separate PRs, as much as possible. Due to the size of the refactor, the sooner
it lands the better for us all.