examples/ct: cope with prechains without roots
Go 1.8 includes commit 07a31bc3da11 ("crypto/x509: don't accept a root
that already appears in a chain"); this changes the behaviour of
Certificate.Verify so that chains will not have redundant roots, i.e.
chains like "Leaf->RootA->RootA" will no longer be produced.

It turns out we were accidentally relying on this behaviour.  If we
fed in Leaf->RootA as input, the old X.509 code would return
both Leaf->RootA and Leaf->RootA->RootA as possible chains, and
the cert_checker code would strip the final root and thus check
Leaf->RootA against either Leaf or Leaf->RootA (and pick the latter).
So even though the code in ValidateChain() claimed to be returning
`chainMinusRoot`, it was still returning a chain with a root at the
end -- and the rest of the code (e.g. building extra_data in
extraDataForChain) actually relied on this.

If the X.509 code were to (correctly) just report Leaf->RootA, we
need to preserve the final root cert to allow for the input chain
including the root.

OTOH if the input chain doesn't include the root, we still need
to accept it.