Add a TODO to NewLogEnv() to simplify code by removing stored context
NewLogOperationManagerForTest() stores the context that is passed to it.
This makes the code in NewLogEnv() more complicated.
If it weren't, a cancelable context could be used for both code paths,
the cancel func could be deferred and all of the checks for the cancel
func being nil could be removed.