admin.Server returns the correct error codes
This PR wires trillian/errors to the most common error sources on admin.Server.
It also adds a method to read the errors.Code from generic errors.

Errors originating from the admin storage layer are still unmmapped, therefore
their code is implied to be Unknown. While other codes would be more appropriate
(Unavailable for Begin errors, Internal for enum-mapping errors, so on), that
should be acceptable for now, as those errors are relatively uncommon.