Make keys.Generator also a keys.SignerFactory
If it is capable of generating a new key, it should also be capable of
retrieving keys. This simplifies testing/mocking. It also reflects the
intended implementation of this interface - a SignerFactory *may*
implement Generator if it can generate keys as well as retrieve them.

As a result of this change, the keys.testonly package is best replaced by
mocks. This also allows some code in sequence_manager_test.go to be
replaced by using these mocks.