Key generation now triggered by a nil tree.PrivateKey
If a tree has a nil PrivateKey, key generation will be attempted.
If the SignerFactory does not implement keys.Generator, then an error
will be returned indicating that PrivateKey must not be nil.

If the private key cannot be found (i.e. SignerFactory.NewSigner() returns
errors.NotFound), key generation will no longer be attempted. An error
will just be returned.