Add a Kubernetes config and Docker image for running a Galera MySQL cluster
This can be used to easily start up a new Galera MySQL cluster running on
a Google Container Cluster.

The MySQL servers will be backed by 100Gi SSDs.
They will be running Percona Server 5.7 and use the XtraDB storage engine.
They will be configured for Trillian testing, so the users will be:
- user="root"; password="zaphod"
- user="test"; password="zaphod"
There will be an empty database named "test".