Interceptor for Trillian RPCs
Add an interceptor for tree-based RPCs (and wire it into servers). RPCs that
don't match the requested tree (non-existing, wrong type, write on frozen tree,
etc) will be blocked at the interceptor layer.

RPCs that do succeed have the requested tree set in context, as to avoid future
queries of the same data within the scope of a request.

The mechanisms introduced here allow for easy addition of new interceptors, for
example auth and rate limiting.