Require PEM password in crypto/keys.NewFromPrivatePEMFile
If an empty password is passed to crypto/keys.NewFromPrivatePEMFile then throw
an error. Also removes the path os.Stat check from cmd/createtree for the
PEM key file and the related invalid path test since this assumes the file
exists on the local machine not the machine that the trillian admin server
is running on (which is where the file is actually loaded off disk).