Allow AdminServer to generate keys for new trees via SignerFactory
A CreateTreeRequest can now specify a key_spec (keyspb.Specification)
that defines how the tree's keys should be generated. The AdminServer
will pass this specification to the SignerFactory's Generate() method,
which will generate a new private key.

For simplicity of implementation, a PrivateKey proto has been added to allow
private keys to be stored along with the tree proto. PEMSignerFactory has been
updated to support this. This is convenient, but less secure than storing the
private key in an encrypted PEM file because storage compromise is all that is
required to obtain the key, whereas encrypted PEMs require both storage
compromise (for access to the password) and access to the PEM files as well.