Address review comments
Remove, rather than replace, the tree's public key in BeginError tests.
Test mismatched and omitted public keys in CreateTree tests.

Because CreateTree now may modify the tree prior to passing it to
AdminTx.CreateTree(), the CreateTree test had to be modified to remove
the assumption that it would not be.