Interface and noop implementation for Trillian quotas
Quotas are meant as a protection against traffic peaks / service DoS. Each
request requires certain tokens, which are replenished over time. If there
aren't enough tokens, requests are denied.

An interface for acquiring / replenishing tokens is defined, along with a Noop
implementation that acts as if it had infinite tokens.