Remove liveness probe for Galera instances
If an SST was underway to replicate the database state, the recipient
Galera instance would appear to be dead, according to this probe, until
the SST completed. This would result in Kubernetes killing the recipient.

TODO: Devise a better liveness probe, perhaps checking whether the file
"/data/mysql/sst_in_progress" exists.