log: force signed root regeneration after interval
For a CT log, RFC 6962 s3.5 specifies that the log should produce a new
STH within the MMD even if there have been no additions to the log.

To allow a CT personality to get the same behavior, force regeneration
of a signed log root after an interval.

This option is implemented in the signer rather than in the personality
because a different user of Trillian-as-a-service may also have the same
requirement that led to RFC 6962 s3.5.

Note that similar functionality was previously removed in commit 0b66fd9
("Remove the code for signing log roots if the last one is too old.")

Addresses issue #575