Return empty list when no leaves are requested. (#613)
The map server currently panics if no map leaves are requested.
This PR adds a test for this and returns an empty list.

Also verify index length.
Indexes in the map should all be the same size.
Currently this size is unspecified. #640
This PR uses the existing implicit constant of 32 bytes and
adjusts the tests to conform to it.