Add --mysql_max_conns flag to Trillian binaries
Controls the maximum number of connections each binary can make to the
MySQL database. Defaults to approximately half of the default limit
imposed by MySQL. This could be increased if using a MySQL cluster, or
if the max_connections setting on the MySQL server has been increased.

This avoids a "too many connections" error under heavy load, so long as
there are only two Trillian binaries running.