Refactor keys package
This will make it easier to exclude key types from the build through the
use of build tags in the future. For example, PKCS#11 key support could be
excluded from the build unless a build tag is used, e.g.
`go build -t pkcs11`. A key benefit is that this would avoid forcing users
to install dependencies for all key types - only those that they include
in their build.