NewMapEnv() for map integration test matrix
This PR creates a MapEnv, similar to LogEnv to support pure go
integration tests for the map.

This supports targeted testing of round-trip AddMapLeaf /
VerifyMapInclusionProof operations for each of the map hashers.

At this time, the TEST_MAP_HASHER is the only one that is known to work.

Also included:
- Added function names along error return paths in the map.