Table driven tests for storage/types.go
Before changing anything, I thougt it would be a good idea to flesh out
the current unit tests to get a good understanding of what the code is
doing.

I've preserved all the existing test vectors, and added some additional
ones.

There are a few scenarios (marked with TODO's) where the current output
seems incorrect, but I could be wrong.  Would love to get clarity on the
correct behavior so I can use this set of unit tests as a bench mark on
whether my changes are consistent.