Make NodeID.PathLenBits a function
Clarify the correct definition of PathLenBits by making it a strict
function of the length of NodeID.Path.

A side affect of this change makes PathLenBits always a multiple of 8.
This PR add checks in a few places to enforce this requirement, which
was previously an assumption present in the codebase.