Replenish tokens for bad requests / duplicate leaves (#741)
Bad requests and, more importantly, duplicate leaves, may be a major source of token leakage for sequencing-based quotas (as tokens are taken when the request goes in, but never replenished via sequencing). Interceptors are changed to now return tokens in those scenarios.

This change introduces the first block of genuine after-handler logic. To allow for easy ports to non-gRPC environments, the concepts of a RequestProcessor and Before/After have been added.

* Define a RequestProcessor and distinguish between Before/After interceptor steps

* Replenish tokens for bad requests / duplicate leaves

* Review comments