Add --generate-key flag to createtree command (#762)
Makes the Trillian admin server generate the private key, rather than
requiring the user to provide one. This may be more secure, depending
on how the server stores the generated key.

It is hard-coded to generate a key using the default parameters.
Users that require more control over the key specification should
communicate with the admin server directly.