Add experimental batched delete/insert dequeuing mode
Adds a new dequeuing mode that uses batched deletes from the Unsequeunced
table and batched inserts into the SequencedLeafData table. This mode is
hidden behind a new build tag 'batched_queue' as it relies on a schema
change and requires further testing to see if it adequately scales in
production environments.