Remove tree deletion states and add "deleted" and "delete_time"
deleted and delete_time are the standard way to represent soft deletion
according to Google's API style guide[1]. Furthermore, it has the added
benefit of "remembering" the correct previous state for undeletion.

Storage and tests have been changed accordingly, even though the actual
Delete / Undelete features are outstanding.

The objective of the PR is not to implement deletion, but to prepare the
existing codebase for it by adding the correct concepts to API, storage
and delete-sensitive methods that already exist.

[1] https://cloud.google.com/apis/design/standard_fields