Change SignedMapRoot metadata to google.protobuf.Any
This leaves the choice of exactly what metadata should be stored in the
signed map head to the map personality.

The previous MapperMetadata proto moves into the examples/ct/ctmapper
as an example of how marshalling should be done.  It is also used in
the existing Trillian map tests as it was before.