Make assertions when storing and retrieving rev's
There are code paths in the hammer where the -1 'not found' value from
rev() was causing unexpected behavior, because -ve rev is valid in
get-leaves.  Change that rev() to panic instead.

Assert that we don't try to store -ve rev values in the hammer state
contents.

Assert that rev increases when we push new content after a set-leaves
call.