Additional request validations for TrillianLogRPCServer (#877)
* Additional request validations for TrillianLogRPCServer

Added additional validations to a few TrillianLogRPCServer methods, most
notably QueueLeaf/QueueLeaves.

Validation tests have been refactored to go through RPC calls, so we can
test both correct wiring and the effectively returned error code. As a
consequence, validation_test.go was removed.

There's a small functionality change: leaf_identity_hash is not a
required parameter anymore. If it's not present, it's assumed to be the
same as merkle_leaf_hash.

* Document leaf_identity_hash change

* Regenerate protos

* Remove redundant usages of LeafIdentityHash

* Review comments: docs and better error messages