GetLeaves@rev 0 is valid since SMR rev 0 exists
Since we now initialize the map with a SignedMapRoot at revision 0
it is now possible to start a Txn on the map at that revision, which
enables readers to GetLeaves on rev 0.  This is a not a behaviour
breaking change as previously to get the most recent leaf revision a
caller would have to specify a -ve revision number in their request.