Make presubmit tests output their glog logs to stderr.
This helps provide more verbose (and meaningful) output when failures
happen, since most of the logging which is useful for debugging uses the
glog library.

Because this adds the `-aslologtostderr` to all unit tests, this change
also makes sure that all unit tests import "github.com/golang/glog".
Otherwise the tests would fail with a "flag provided but not defined".