Use gRPC errors throughout
As of grpc/grpc-go#1156, the gRPC errors package is a separate
self-contained pair of packages:
 - google.golang.org/grpc/status for manipulating Status / errors
 - google.golang.org/grpc/codes for error code values.

This means that pulling in gRPC errors no longer pulls in the whole
of gRPC, which removes one of the main reasons why we had our own
Trillian errors package.

So shift to use gRPC errors throughout the codebase.