Drop use of SQLite (#1064)
This partly-reverts commit 4023e5ab7de5 ("Use sqlite3 as the default
driver for unit tests (#886)"), removing SQLite as a storage option.

In practice, having a default DB that doesn't support much of the
Trillian functionality (e.g. none of the map tests, none of the
integration tests) has inconvenienced core developers.

However, to allow the core `git clone <repo> ; go test ./...`
scenario to still work without setup dependencies, all DB-using tests
explicitly check for MySQL availability and are skipped if MySQL
is not present.  A single deliberate test failure is emitted in this
situation as a reminder that many tests are being skipped.