[LogVerifier] Introduce PrefixHashFromInclusionProof
The new method calculates root hash of a smaller tree which is a prefix
of a bigger tree based on an inclusion proof of the boundary entry to
the bigger tree. If the passed in root hash of the bigger tree is
trusted, then the result of this function is likewise trusted.

This function can be used, for example, in monitoring applications to
verify intermediate root hashes of a log that is being copied. If, say,
a log publishes its root hash for 1 million entries, it is not necessary
to download the lot to verify only a prefix of 100 entries. Instead,
PrefixHashFromInclusionProof can be utilized.