Remove vendored tools for CI.
In preparation to vendor the dependencies of Trillian itself with `dep`,
this change removes the existing contents of the vendor directory, and
updates the .travis.yml to instead check out specific versions of the
commands that were in the vendor directory (etcd, etcdcli, and
protoc-gen-grpc-gateway) , matching the previously checked-in versions.

This is because vendoring with `dep`, which is approximately
representative of the direction Go as a whole is moving in, is
incompatible with vendoring arbitrary commands.