Retry gometalinter when it times out.
Increasing the deadline did not seem to decrease the failure rate due to
timing out. Perhaps this is because some other factors are at play
causing gometalinter to hang.

Therefore, this change attempts to instead retry the failed command
(using a reasonable timeout) rather than only run it once with a large
timeout.