Remove the updateRootIfNewer method
This is because it's not immediately obvious from the method signatures
that this is a security critical method and should only be called after
a successful verification of the new root, and not just any root with a
newer timestamp.

By removing the method, we can avoid potential misuse, which can
compromise the security guarantees that the log_client provides.