Improve the logic around updating the trusted root (#1225)
* Add a LogClient method that returns a copy of the latest trusted root.

* Return a `*types.LogRootV1` from `GetRoot` for consistency.

Most other methods use pointers to the struct (both for arguments and
return values).

* Improve methods for updating the trusted root.

- Add an `updateRootIfNewer` private method, which is used to update the
  trusted root by other methods, and only performs the update if the
  updated tree is more recent and larger than the current trusted one.

  By using this helper method, other methods avoid holding on to the
  `rootLock` for longer than necessary. For example, before this update
  `UpdateRoot` would hold on to `rootLock` while calling `getLatestRoot`
  which makes multiple RPC calls to get the new root and fetch its
  consistency proof.

- Remove the `waitForTreeSize` argument from `WaitForRootUpdate`. This
  argument makes the function harder to use and adds complexity to it.

  This complexity also introduced the following subtle bug:

  If `WaitForRootUpdate` was called with
  `waitForTreeSize >= current size + 2`, then the root might be updated
  (ex. to `current size + 1`) and then an error encountered during the
  next update. When this happens, the method returns with an error and
  with `nil` for the new root (incorrectly indicating no root update).

* Update the documentation for `WaitForInclusion`.

Also, remove an unused value.

* `VerifyInclusion` and `GetAndVerifyInclusionAtIndex` to not update.

Those two methods don't need to update the root, so this change makes
them use the current one. If a user does require a trusted root update,
they can do that by calling `UpdateRoot` or `WaitForRootUpdate` manually
before.

* Update the documentation for `AddSequencedLeafAndWait` and `AddLeaf`.

Clarify that those methods update the trusted root under the hood.

* Call `UpdateRoot` in `TestVerifyInclusion`

* Make sure to wait for the root update in tests

* Remove root updates from TestVerifyInclusion tests.

This is because `addSequencedLeaves` already updates the root itself.

* Rename getLatestRoot to getAndVerifyLatestRoot

* Remove the updateRootIfNewer method

This is because it's not immediately obvious from the method signatures
that this is a security critical method and should only be called after
a successful verification of the new root, and not just any root with a
newer timestamp.

By removing the method, we can avoid potential misuse, which can
compromise the security guarantees that the log_client provides.

* Fix a race condition in `UpdateRoot`

Multiple `UpdateRoot`s running concurrently can result in an invalid
update if the Trillian log was forked and each update was given a
different branch of the fork.