Stop the MySQL storage provider from losing its mysql.OpenDB error
Currently, if `server.NewStorageProvider("mysql", nil)` is called twice
with a bad MySQL URL, the first time, it will fail properly returning a
non-nil error. However, this error is lost afterwards and the second
time it will silently fail and return a nil StorageProvider.

This change fixes that by making the MySQL storage provider hold on to
any initialization errors it encounters, and returning them for every
new storage provider request.