Make sure to always set the Go MySQL driver's `parseTime` flag
The ability to parse MySQL DATE and DATETIME fields as time.Time objects
is currently behind a DSN-activated flag that's disabled by default.

Since the logmonitor makes use of this feature, and because it's an
implementation detail for Go's MySQL driver, this change always enables
the flag on MySQL connection strings provided by the user.