docker-compose: Run log integration test against Trillian-in-Docker
To make the integration test work with Docker, it now accepts an
optional "log server address" parameter. If this parameter is provided,
the integration test will not start its own test instance of Trillian.
Instead, it will communicate with an existing deployment via the log
server listening at the given address.