mapserver: deal with non-existent leaves
Firstly, when a key has never been set, the hash value for the
associated leaf is effectively HashEmpty(index, 0), and this
is represented on the API as a nil value.

Secondly, when a client removes the value for a key/index, it does so by
setting a zero length value. Such requests should not be silently
dropped, but should be stored. The hash value for this leaf afterwards
is HashLeaf(index, nil) not HashEmpty(index, 0).