docker-compose: Only test on Travis if PKCS#11 is disabled
PKCS#11 integration testing is not yet supported, because it is not
possible to install softhsm into the Trillian containers (they use
"distroless" images, which lack package managers). It should be possible
to run softhsm in a separate container and use a PKCS#11 proxy to expose
it, but this is non-trivial so is left for the future.