Only retry retryable errors.
Only retry errors that are retryable. Stop the existing
behavior of retrying unknown errors as this could be dangerous.

Existing callers of Retry() have to return nil to get Retry() to exit,
and lose err in the process.

Add an option for callers of Retry() to explicitly declare what errors
they are expecting to be retryable if they deviate from standard
behavior.