Prevent signer runs over inactive logs (#1319)
Signer tries to process all logs that it holds mastership for. This
change refactors it so that only active (i.e. not deleted or FROZEN)
logs of those are taken into consideration. This prevents signer
producing a sustaining stream of errors saying that it's unable to
sequence logs that are in fact inactive.

+ Clean up mastership status logging, add comments