Add db_server Docker image to cloudbuild_tag.yaml
This will cause it to be built and pushed to our GCP Container Registry
automatically every time a new Git tag is created. The db_server image
is the simplest way to run Trillian on Kubernetes, because it provides a
properly-configured MySQL database for Trillian to use. However, a
production setup of Trillian should use a more scalable, reliable solution,
such as Cloud SQL or Cloud Spanner.