election2: Add rigorous Election testing (#1386)
This change adds a test-only Election implementation and tests for it.
It also introduces a Decorator that allows injecting errors into
Election implementations during testing, and controlling how long Await
method is blocked. The decorator and the bulk of the tests code is
re-used by both testonly.Election and etcd.Election implementations.