Remove warning logging from log/sequencer.go (#1418)
* Remove warning logging from log/sequencer.go

In favour of including the same information in error messages. The caller
can choose to log the error if they wish. This avoids misleading warning
messages in logs when the storage implementation retries transaction
errors. While storage can swallow the errors when a retry succeeds, it
can't cancel the logging.