Run docker-compose integration test as separate Travis job
This avoids slowing down the jobs that are required to complete in order
to merge a pull request, without entirely removing this test. If a pull
request author waits the extra 5 minutes or so required by this test,
then they can have confidence that their change doesn't break our Docker
images or docker-compose configs.