api: document empty proof behaviour
Commit 9a5dc6223bab0 ("Example of dealing with server skew. (#995)")
modified the Trillian Log server behaviour for when a proof is
requested that is beyond the currently known tree size; in this case
the returned proof is empty, and the associated log root includes
the known tree size information.

Document this behaviour so that clients can be aware of it.

Also regenerate all top-level files derived from .proto inputs.