Add flag to limit max idle database connections for mysql
We've been using Trillian in a serverless (function as a service) context and
using Amazon's Aurora serverless database as a backend. Aurora will scale down
the provisioned databases based upon open connections and CPU usage.

Making this a configurable flag allows us to set the idle pool to 0 and for
Aurora to scale down to 0 running databases.

Go currently defaults to pooling 2 idle connections:
  https://golang.org/pkg/database/sql/#DB.SetMaxIdleConns

Defaults to -1 which uses the Go default value (2 at time of commit).

Co-authored-by: Emily Mattiussi <emily@projectsbyif.com>