More efficient Equivalent method for storage.NodeID
This was converting them to binary strings and somewhat obfuscating the
underlying comparison by the fact that the Bit method counts from the
right.

This version does the comparison without memory allocation on the raw
path bytes.

This method has been used in a few places where performance matters a
bit more, including comparing results in the map hashing code.