Rollback if QueueLeaves runs on uninitialized log
Not sure how common this is to encounter but if QueueLeaves is called on an
uninitialized log the tx isn't closed. This makes the next call trying to
acquire the mutex to deadlock.

Example of the issue is here
  https://gist.github.com/davbo/99a01ce5ffe482b55698e74a10f94af6

This change rolls back on all errors, not just uninitialized logs.

Co-authored-by: Emily Mattiussi <emily@projectsbyif.com>