Fixes for transaction leaks in RPC server and storage level operations (#1480)
* Fixes for transaction leaks in RPC server and storage level operations

QueueLeaves / AddSequenced leaves (Storage Level API) in memory / mysql
storage could leak a transaction in 2 ways - on an ErrTreeNeedsInit or
if the operation they call on the transaction failed. Transactions are
created inside these methods.

The log RPC server could leak via the above for write methods or in a
similar manner for read RPCs by not closing a TX when receiving an
ErrTreeNeedsInit from the initial Snapshot() call.

Add tests. Update some tests to ensure Close() is called for errors by
removing AnyTimes().

* Update changelog

* Close the tx if non nil for any error.

* Fix changelog nits

* Do the same fix for map_rpc_server read RPC methods.

The `SetLeaves` RPC doesn't have the same problem as the log as it uses
`ReadWriteTransaction` directly, which closes any non-nil TX.

* Update changelog to mention the map server