Make storage.Suffix immutable and share instances for 8 bit values. (#1507)
* Start making Suffix immutable, add accessors, unexport values

Fix up tests. Add a test that modifying Path() results does not change
the Suffix.

* Create a reused empty suffix var.

Move to a constructor for Suffix. Make the string rep a field and
compute it once when constructed. We expect it to be called at least
once on all Suffix instances more or less wave hands etc. etc.

* Make the Split API use pointer to Suffix

* Maintain a cache of 8 bit Suffix values.

Return them from NewSuffix and ParseSuffix as appropriate.

* Fix test

* Additional tests that the cache works and users can't corrupt it.

* Review fixes

* Fix typo. Add more bit lengths to serialize tests.

* Fix up a suffix test that was weaker than appeared

And add more test cases.

* Make the Suffix caching work with depths 1-8

Instead of just depth 8. Update tests.