Experimental LogHasher API change for performance.
Allows hash buffers to be reused at the cost of thread safety and the
need to make copies in some places. Apply this to the compact merkle
tree to demonstrate the gains.

Performance improvement is noticeable

```
BenchmarkAddLeaf-12    	  200000	     15023 ns/op
BenchmarkAddLeaf-12    	  200000	     11392 ns/op
```

Also memory profile showed ~20% reduction in memory allocation for the
benchmark run.