Faster version of NodeID.String() for use by sparse_merkle_tree (#1557)
* Faster version of NodeID.String()

For use in mapkeys etc. Add it to the sparse_merkle_tree code where it's
used in proofs. Profiling showed this using over 3% of CPU time in a
maphammer run. This version is 10x faster.

```
BenchmarkString-12    	 1000000	      2989 ns/op
BenchmarkAsKey-12     	 5000000	       307 ns/op

```

The Log sequencer does a similar thing but I've left that alone for now
so this change only affects the map.

* Make it look like we use the bench result to fix the linter.