Overwrite buildCtx.BuildTags instead of appending to it
Since buildCtx is a shallow copy of build.Default, appending to
buildCtx.BuildTags will modify the underlying array shared with
build.Default. This means other code that appended to
build.Default.BuildTags later would overwrite the elements added via
buildCtx. While this is an unlikely scenario, it would be confusing to
debug so we should simply avoid it now.

This change assumes that any build tags set on build.Default should
not be preserved. Since we have a flag for setting build tags, this
seems reasonable.