compact.Tree: Compute root hash on demand
This change removes root hash field from compact.Tree so it is no longer
necessary to maintain it after each AddLeafHash operation. A side effect
of this is that CurrentRoot now can return an error. We make sure that
the error is handled in the client code.

This commit prepares us to remove some unnecessary computation when many
AddLeafHash calls are done in a row. Ephemeral nodes along the right
border of the tree, and the root hash will be computed only once.