CONIKS, hstar2 and SMT Writer performance improvements (#1575)
* Make coniks hasher accumulate into buffers and write them to hasher

It's about 10% faster for reasons. Avoid reallocating in maskIndex() by
writing directly to the buffer.

* Make hstar2 use pointers to leaf hashes throughout

Update sparse tree code + tests to use them.

* Remove some slice copying in hstar2 + resolve TODOs

Nothing was modifying the copies.

* Fix leaf hash usage in map subtree cache

* Fix bug that makeIndex was writing without padding.

* Save a few bytes of ram. Allow for hashers > 256 bits in maskIndex

* Fixes, comment doc updates and extra tests on a 512 bit hasher.

* Rename maskIndex -> writeMaskedIndex

* Fixup off by one problem

* Fix incorrect error message.

Depth can be zero in the precondition checks but the message said it
must be > 0.

* Add a crosscheck test on writeMaskedIndex.

* Preallocate buffers of suitable size.

Looks like it might be worth it from bencharking.