licenses.Libraries() returns the Library containing the given pkg
Previously, it would not return a Library for the provided Go package,
or for any packages that were part of that Library. This changes its
behaviour to do so. This is an improvement because command output will
now be comprehensive - it includes all license information, including
the license that covers the target package (no just dependencies).