Format storage/mysql/schema/storage.sql using `skeema lint`
I had to manually replace "CREATE [UNIQUE] INDEX" statements with
"[UNIQUE] KEY" options within the "CREATE TABLE" statements, because
Skeema doesn't support the former.

I've made the comments on columns/indices into actual column/index
comments, because Skeema deletes comments within "CREATE TABLE" statements.
This makes them a little harder to read in the file, but will surface them
through tools that read the live database schema. A word of warning is
that Skeema will delete and recreate the primary key of the Subtree table
in order to add a comment to it, which could take a while if the table
has a lot of rows.

Skeema deletes the "IF NOT EXISTS" part of the "CREATE TABLE" statements
because it doesn't support them. This should be fine if we switch to using
Skeema, because it handles this functionality (creating or altering tables
as necessary).