Avoid an unnecessary Split() in preload()
The prefix can be used directly from the NodeID path slice, it's only
used to create a hash table key. The code already effectively assumes
that the prefix is a multiple of 8 bits. For one thing the subtree
strata store prefix lengths in terms of bytes not bits.

Not sure how good the tests are for this code. I'll investigate.