New API service for writing to the Trillian Map (#1655)
Separating the Verifiable Map along the concerns of reading vs writing leads to a cleaner API.

This change simply adds a new API service for writing to the map. Future changes will remove the methods from the read API when the service has been implemented and clients have been migrated.