subtree_cache: Lock reads and writes separately (#1798)
This change removes a mutex lock in SubtreeCache from the place where database
reads are. Now only the in-memory reads and writes are under lock. This allows
double-reads from DB if multiple parallel reads of the same node hash occur,
but, having looked at the call sites, it never happens. The cache is used on
per-transaction basis, and within one transaction all reads are either done in
one batch, or in batches isolated from each other.

This speeds up Map integration tests by 5-10%.