Factor out treeLayout type from SubtreeCache (#1837)
This change moves the subtree strata calculus into a separate treeLayout type.
This is the first step towards having cleaner abstraction levels around
SubtreeCache. As an extra small optimization, this type can have 2 global
instances (for log and for map), and be reused everywhere instead of being
instantiated whenever a new transaction is started.