Reduce the min/max leaves in the hammer test
This load causes failures on multiple versions of MySQL. This has been going on for a while and doesn't seem to be the result of any recent changes, but has come to light due to mariadb being force upgraded on our dev machines.

This is a temporary workaround while we determine: a) what versions of MySQL we want to support, and b) what loads are reasonable to hit it with.